requirements.txt
crewai==0.80.0
crewai-tools==0.17.0
google-generativeai==0.8.3
python-dotenv==1.0.1
requests==2.32.3
langchain-google-genai==2.0.7
